move io.apo file to training
